Add a new tests, update READMEN and REQUIREMNTS files.

+++ migrations/sqls/20221212155706-orders-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     order_status BOOLEAN,
-    user_id bigint REFERENCES users(id)
+    user_id int REFERENCES users(id)
 );
