+++ migrations/sqls/20221226005724-order-product-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE order_product (
     id SERIAL PRIMARY KEY,
     quantity integer,
-    product_id bigint REFERENCES products(id),
-    order_id bigint REFERENCES orders(id)
+    product_id int REFERENCES products(id),
+    order_id int REFERENCES orders(id)
 )
